Introduce ReplicationProcessor/Stopper
ReplicationProcessor combines a ReplicationFilter and a
ReplicationStopper. A ReplicationStopper given an event to be
replicated decides if the replication shall stop on this event. This
decision can be taken on source (e.g. RFC) or target side (e.g.
because of schema evolution).

The ReplicationProcessor applies filtering and stop-decisions on a
sequence of events and returns the remaining events to be replicated
(and a modified progress, if required) or a reason why the
replication was stopped.